+++ objc2-foundation/src/dictionary.rs
@@ -5,7 +5,7 @@ use core::ops::Index;
 use core::ptr::{self, NonNull};
 
 use objc2::msg_send;
-use objc2::rc::{Id, Owned, Ownership, Shared, SliceId};
+use objc2::rc::{DefaultId, Id, Owned, Ownership, Shared, SliceId};
 
 use super::{INSCopying, INSFastEnumeration, INSObject, NSArray, NSEnumerator};
 
@@ -153,6 +153,15 @@ impl<K: INSObject, V: INSObject> NSDictionary<K, V> {
     unsafe_def_fn!(pub fn new -> Shared);
 }
 
+impl<K: INSObject, V: INSObject> DefaultId for NSDictionary<K, V> {
+    type Ownership = Shared;
+
+    #[inline]
+    fn default_id() -> Id<Self, Self::Ownership> {
+        Self::new()
+    }
+}
+
 unsafe impl<K: INSObject, V: INSObject> INSDictionary for NSDictionary<K, V> {
     type Key = K;
     type Value = V;
@@ -166,7 +175,7 @@ unsafe impl<K: INSObject, V: INSObject> INSFastEnumeration for NSDictionary<K, V
 impl<'a, K: INSObject, V: INSObject> Index<&'a K> for NSDictionary<K, V> {
     type Output = V;
 
-    fn index(&self, index: &K) -> &V {
+    fn index<'s>(&'s self, index: &'a K) -> &'s V {
         self.get(index).unwrap()
     }
 }