+++ objc2/src/rc/mod.rs
@@ -1,30 +1,59 @@
 //! Utilities for reference counting Objective-C objects.
 //!
-//! The utilities of the `rc` module provide ARC-like semantics for working
-//! with Objective-C's reference counted objects in Rust.
+//! These utilities in this module provide ARC-like semantics for working with
+//! Objective-C's reference counted objects.
 //!
 //! A smart pointer [`Id`] is provided to ensure that Objective-C objects are
-//! retained and released at the proper times.
+//! retained and released when created and dropped, respectively.
 //!
 //! To enforce aliasing rules, an `Id` can be either owned or shared; if it is
 //! owned, meaning the `Id` is the only reference to the object, it can be
 //! mutably dereferenced. An owned `Id` can be downgraded to a shared `Id`
 //! which can be cloned to allow multiple references.
 //!
-//! Weak references may be created using the [`WeakId`] struct.
+//! Weak references may be created using the [`WeakId`] struct; these will not
+//! retain the object, but they can upgraded to an `Id` in a manner that
+//! safely fails if the object has been deallocated.
 //!
 //! See [the clang documentation][clang-arc] and [the Apple article on memory
-//! management][mem-mgmt] (similar document exists [for Core Foundation][mem-cf])
+//! management][mem-mgmt] (similar document exists [for Core Foundation][cf])
 //! for more information on automatic and manual reference counting.
 //!
 //! It can also be useful to [enable Malloc Debugging][mem-debug] if you're trying
 //! to figure out if/where your application has memory errors and leaks.
 //!
-//!
 //! [clang-arc]: https://clang.llvm.org/docs/AutomaticReferenceCounting.html
 //! [mem-mgmt]: https://developer.apple.com/library/archive/documentation/Cocoa/Conceptual/MemoryMgmt/Articles/MemoryMgmt.html
-//! [mem-cf]: https://developer.apple.com/library/archive/documentation/CoreFoundation/Conceptual/CFMemoryMgmt/CFMemoryMgmt.html
+//! [cf]: https://developer.apple.com/library/archive/documentation/CoreFoundation/Conceptual/CFMemoryMgmt/CFMemoryMgmt.html
 //! [mem-debug]: https://developer.apple.com/library/archive/documentation/Performance/Conceptual/ManagingMemory/Articles/MallocDebug.html
+//!
+//!
+//! ## Example
+//!
+#![cfg_attr(apple, doc = "```")]
+#![cfg_attr(not(apple), doc = "```no_run")]
+//! use objc2::{class, msg_send};
+//! use objc2::rc::{autoreleasepool, Id, Shared, WeakId};
+//! use objc2::runtime::Object;
+//!
+//! // Id will release the object when dropped
+//! let obj: Id<Object, Shared> = unsafe {
+//!     Id::new(msg_send![class!(NSObject), new])
+//! };
+//!
+//! // Cloning retains the object an additional time
+//! let cloned = obj.clone();
+//! autoreleasepool(|pool| {
+//!     // Autorelease consumes the Id, but won't
+//!     // actually release until the end of an autoreleasepool
+//!     let obj_ref: &Object = cloned.autorelease(pool);
+//! });
+//!
+//! // Weak references won't retain the object
+//! let weak = WeakId::new(&obj);
+//! drop(obj);
+//! assert!(weak.load().is_none());
+//! ```
 
 mod autorelease;
 mod id;