+++ objc2-encode/CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased - YYYY-MM-DD
 
+### Added
+* Implement `Hash` for `Encoding`.
+
+### Changed
+* Improved documentation.
+
 
 ## 2.0.0-beta.1 - 2021-12-22